+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.1.5
+* Split the test project;
+* Use `Spider#addUrls` to add urls to urls queue, instead of using `Spider#urls.enqueue()`.
+
 ## 0.1.4
 * Change handler arguments;
 * Retry for requesting failure.